Start fleshing out src/eigen.d.ts
